fix: Complex type restrictions now allowed
"type_your_inputvalue: parse[name]" is allowed and will be passed over
if the tag parses. This still does not allow specific values to be
checked for like type_list: parse[name].contains[chemical_datas]. Even
if the name isnt in the list, as long as the tag parses, this will pass
over.