added New commands: save_database, load_databases, icons, console_mode
save, load databases is for force saving loading yaml files uses with
'yaml load' 'yaml save' scripts.

icons outputs all of the [icons] available for use with the parser.

console_mode is a wip for redirecting output to the players chat for
monitoring the console and input as the console.